experiments/embedding-search-quality/requirements.txt
google-genai>=1.0
markitdown[all]>=0.1
lxml>=5.0
python-docx>=1.1
python-pptx>=1.0
numpy>=1.26
